# Experiment: Implement tactics using exceptions and inline bind.