--- a/src/extraction/FStarC.Extraction.ML.Term.fst
+++ b/src/extraction/FStarC.Extraction.ML.Term.fst
@@ -1624,6 +1624,16 @@ and term_as_mlexpr' (g:uenv) (top:term) : (mlexpr & e_tag & mlty) =
                args |> should_apply_to_match_branches ->
           args |> apply_to_match_branches head |> term_as_mlexpr g
 
+        (* HACK HACK HACK HACK HACK
+           Ideally we'd put inline_for_extraction on tac_bind and lift_div_tac_lid,
+           but that causes norm_reify to blow up spectacularly (with unbound variables).
+           Therefore we unfold them here manually, this blows up more rarely. *)
+        | Tm_app {hd={n=Tm_uinst({n=Tm_fvar fv},_)}; args}
+        | Tm_app {hd={n=Tm_fvar fv}; args}
+            when S.fv_eq_lid fv PC.tac_bind_lid || S.fv_eq_lid fv PC.lift_div_tac_lid ->
+          let t = N.normalize [Env.UnfoldOnly [PC.tac_bind_lid; PC.lift_div_tac_lid]; Env.HNF; Env.Beta; Env.EraseUniverses; Env.AllowUnboundUniverses] (tcenv_of_uenv g) t in
+          term_as_mlexpr g t
+
         (* A regular application. *)
         | Tm_app {hd=head; args} ->
           let is_total rc =