--- a/tests/error-messages/Bug1918.fst.output.expected
+++ b/tests/error-messages/Bug1918.fst.output.expected
@@ -1,5 +1,5 @@
 * Info at Bug1918.fst(11,13-11,14):
   - Expected failure:
+  - Tactic failed
   - Could not solve typeclass constraint `Bug1918.mon`
-  - See also FStar.Tactics.Typeclasses.fst(395,6-398,7)