--- a/tests/error-messages/Bug2899.fst.output.expected
+++ b/tests/error-messages/Bug2899.fst.output.expected
@@ -5,7 +5,7 @@
   - Reduction stopped at:
       FStar.Stubs.Tactics.Result.Success (Prims.admit ()) "(((proofstate)))"
   - The term contains an `admit`, which will not reduce. Did you mean `tadmit()`?
-  - See also FStar.Tactics.Effect.fsti(215,48-215,58)
+  - See also FStar.Tactics.Effect.fsti(200,48-200,58)
 
 * Info at Bug2899.fst(10,12-10,18):
   - Expected failure:
@@ -13,7 +13,7 @@
   - Tactic got stuck!
   - Reduction stopped at: Prims.admit ()
   - The term contains an `admit`, which will not reduce. Did you mean `tadmit()`?
-  - See also FStar.Tactics.Effect.fsti(215,48-215,58)
+  - See also FStar.Tactics.Effect.fsti(200,48-200,58)
 
 * Info at Bug2899.fst(13,12-13,18):
   - Expected failure:
@@ -22,12 +22,12 @@
   - Reduction stopped at:
       FStar.Stubs.Tactics.V2.Builtins.dump (Prims.admit ()) "(((proofstate)))"
   - The term contains an `admit`, which will not reduce. Did you mean `tadmit()`?
-  - See also FStar.Tactics.Effect.fsti(215,48-215,58)
+  - See also FStar.Tactics.Effect.fsti(200,48-200,58)
 
 * Info at Bug2899.fst(17,50-17,66):
   - Expected failure:
   - Tactic failed
   - Tactic got stuck!
   - Reduction stopped at: reify (tac ()) "(((proofstate)))"
-  - See also FStar.Tactics.Effect.fsti(215,48-215,58)
+  - See also FStar.Tactics.Effect.fsti(200,48-200,58)