--- a/tests/error-messages/MetaRanges.fst.output.expected
+++ b/tests/error-messages/MetaRanges.fst.output.expected
@@ -1,30 +1,33 @@
 * Info at MetaRanges.fst(9,15-9,16):
   - Expected failure:
   - Tactic failed
-  - See also FStar.Tactics.V2.Derived.fst(74,4-74,35)
+  - See also MetaRanges.fst(9,8-9,13)
 
 * Info at MetaRanges.fst(16,10-16,12):
   - Expected failure:
+  - Tactic failed
   - Could not solve typeclass constraint `MetaRanges.deq Prims.int`
-  - See also FStar.Tactics.Typeclasses.fst(395,6-398,7)
 
 * Info at MetaRanges.fst(19,9-19,11):
   - Expected failure:
+  - Tactic failed
   - Could not solve typeclass constraint `MetaRanges.deq Prims.int`
-  - See also FStar.Tactics.Typeclasses.fst(395,6-398,7)
 
 * Info at MetaRanges.fst(23,9-23,15):
   - Expected failure:
+  - Tactic failed
   - Could not solve typeclass constraint `MetaRanges.deq Prims.int`
-  - See also FStar.Tactics.Typeclasses.fst(395,6-398,7)
+  - See also MetaRanges.fst(23,9-23,11)
 
 * Info at MetaRanges.fst(27,17-27,18):
   - Expected failure:
+  - Tactic failed
   - Could not solve typeclass constraint `MetaRanges.deq Prims.int`
-  - See also FStar.Tactics.Typeclasses.fst(395,6-398,7)
+  - See also MetaRanges.fst(27,9-27,11)
 
 * Info at MetaRanges.fst(31,17-31,22):
   - Expected failure:
+  - Tactic failed
   - Could not solve typeclass constraint `MetaRanges.deq Prims.int`
-  - See also FStar.Tactics.Typeclasses.fst(395,6-398,7)
+  - See also MetaRanges.fst(3,33-3,38)