--- a/tests/tactics/FailFlow.fst
+++ b/tests/tactics/FailFlow.fst
@@ -27,8 +27,6 @@ let print_test () : Tac unit =
     print "not failing";
     assert False
 
-(* None of these succeed (as in: return Success within the monad) *)
-[@@expect_failure]
 let s_fail_flow () : TacS unit =
     fail "failing";
     assert False