--- a/ulib/FStar.Tactics.Typeclasses.fst
+++ b/ulib/FStar.Tactics.Typeclasses.fst
@@ -104,7 +104,7 @@ for them and report proper errors. *)
 exception Next
 
 let skip #a (st:st_t) (s : string)
-  : TAC a (fun ps post -> post (FStar.Stubs.Tactics.Result.Failed Next ps))
+  : TacH a (fun _ -> True) (fun ps post -> False)
   = if st.dbg then
       print ("skip: " ^ s);
     raise Next