--- a/ulib/ml/plugin/FStarC_Tactics_V1_Builtins.ml
+++ b/ulib/ml/plugin/FStarC_Tactics_V1_Builtins.ml
@@ -127,7 +127,7 @@ let free_uvars              = from_tac_1 B.free_uvars
 
 (* The handlers need to "embed" their argument. *)
 let catch   (t: unit -> 'a __tac): ((exn, 'a) either) __tac = from_tac_1 TM.catch   (to_tac_0 (t ()))
-let recover (t: unit -> 'a __tac): ((exn, 'a) either) __tac = from_tac_1 TM.recover (to_tac_0 (t ()))
+let raise_core = from_tac_1 TM.traise
 
 let ctrl_rewrite
     (d : direction)