As expected, this reduces Pulse runtime significantly: compiling the Pulse library is 7% faster with this change.  Even compiling the Pulse checker is 5% faster.

 - `make -C pulse plugin ADMIT=1`: 1289s -> 1223s
 - `make -C pulse lib-pulse ADMIT=1`: 611s -> 569s
 - `du -sh pulse/build/checker.ml`: 8.5M -> 5.4M
 - `du -sh pulse/build/ocaml/installed/lib/pulse/pulse.cmxs`: 16M -> 15M

Unfortunately putting inline_for_extraction on tac_bind crashes extraction in Pulse.  I added a creative workaround but that still requires changes in Pulse.

The corresponding Pulse changes are here: https://github.com/FStarLang/pulse/compare/gebner_exntactic?expand=1

It would also be interesting to benchmark implementing tactics as `ref proofstate -> ML 'a`.