@gebner 2025-09-17
Closing in favor of #3999.